+++ src/contexts/action.ts
@@ -3,6 +3,15 @@ import { createContext, useContext } from "react";
 
 interface Action {
 	action: FC | Element;
+	/**
+	 * Set the action to null to hide it from the action bar
+	 *
+	 * ## Example
+	 * ```tsx
+	 * const { setAction } = useAction();
+	 * setAction(null);
+	 * ```
+	 */
 	setAction: Dispatch<SetStateAction<FC | Element>>;
 }