+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 -   feat: Add lottery app
 -   feat: Add hotpot timer app
+-   feat: Rename DrawLink to Mono Card
+-   feat: Use Feishu to send feedback
+-   fix: Hide action bar in feedback page
 
 ## 2.3.0 (2025-01-12)